+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Release 0.11.0 (unreleased)
 * Use CycloneDX schema version 1.6 (#542)
 * Add security policy (#784)
 * Add provenance / release attestation to pypi package (#784)
+* Support multiple licenses per project (#788)
 
 Release 0.10.0 (released 2025-03-12)
 ====================================